requirements.txt
# Core Orchestration
prefect>=2.14.0,<3.0.0

# Cloud SDKs
boto3>=1.34.0                    # AWS SageMaker
google-cloud-aiplatform>=1.38.0  # GCP Vertex AI
azure-ai-ml>=1.12.0             # Azure ML
azure-identity>=1.15.0          # Azure auth

# Data & ML
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0

# Web Dashboard
flask>=3.0.0
flask-cors>=4.0.0

# Config & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Utilities
httpx>=0.25.0
structlog>=23.2.0
rich>=13.7.0
tenacity>=8.2.0
click>=8.1.0
apscheduler>=3.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
